pyproject.toml
[project]
name = "chaosagent"
version = "0.1.0"
description = "ChaosAgent v2: Unreliable Tools Resilience Environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
    "openenv-core[core]>=0.2.2",
    "openai>=1.90.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["models", "client", "inference", "demo"]
packages = ["server", "server.tools"]

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["\\.venv"]
